リストと詳細のデータ連携

+++ backend/src/main/kotlin/com/example/bedrock/repository/ProductRepository.kt
@@ -3,6 +3,6 @@ package com.example.bedrock.repository
 import org.springframework.data.jpa.repository.JpaRepository
 
 
-interface ProductRepository : JpaRepository<Product, Int> {
+interface ProductRepository : JpaRepository<Product, Long> {
 
 }
